creating a standard format time library and fixing what I hope is the last major hurdle for remote browser testing
