+++ lib/terminal/utilities/readme.md
@@ -13,5 +13,6 @@ The various utility libraries for the terminal instance of the application.
 * **[log.ts](log.ts)**                                       - A log utility for displaying multiple lines of text to the terminal.
 * **[readFile.ts](readFile.ts)**                             - A utility to read files as text, if text, or as binary, if binary.
 * **[readStorage.ts](readStorage.ts)**                       - Reads all the storage files and returns a data structure to a callback
+* **[time.ts](time.ts)**                                     - Generates a timestamp in format: "[HH:mm:ss:mil] message".
 * **[vars.ts](vars.ts)**                                     - Globally available variables for the terminal utility.
 * **[wrapIt.ts](wrapIt.ts)**                                 - A tool to perform word wrap when printing text to the shell.
